requirements.txt
asgiref==3.8.1
celery==5.4.0
Django==5.0.6
django-celery-beat==2.6.0
django-ninja==1.1.0
email-validator==2.1.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
pydantic==2.7.4
redis==5.0.7
sqlparse==0.5.0
whitenoise==6.7.0
pytest
pytest-django
black==24.3.0
pre-commit==3.6.0